Extract functions which generate a student or tutorial for tests.
